flashlight implemented
