+++ Assets/Prefabs/AlienTest.prefab
@@ -15,7 +15,7 @@ GameObject:
   - component: {fileID: 1212212924926659335}
   - component: {fileID: 6209905677077487141}
   m_Layer: 0
-  m_Name: Alien
+  m_Name: AlienTest
   m_TagString: Alien
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -65,7 +65,7 @@ MeshRenderer:
   m_RenderingLayerMask: 1
   m_RendererPriority: 0
   m_Materials:
-  - {fileID: 2100000, guid: 31321ba15b8f8eb4c954353edc038b1d, type: 2}
+  - {fileID: 2100000, guid: 165359c3d93954745a19377cdfb6616d, type: 2}
   m_StaticBatchInfo:
     firstSubMesh: 0
     subMeshCount: 0
