+++ Assets/_Scenes/FlashLight_Test_Scene.unity
@@ -1213,7 +1213,7 @@ MeshRenderer:
   m_RenderingLayerMask: 1
   m_RendererPriority: 0
   m_Materials:
-  - {fileID: 2100000, guid: 31321ba15b8f8eb4c954353edc038b1d, type: 2}
+  - {fileID: 2100000, guid: a7dd19c1fa9868845969d9f74f93e4da, type: 2}
   m_StaticBatchInfo:
     firstSubMesh: 0
     subMeshCount: 0
@@ -1252,8 +1252,8 @@ Transform:
   m_GameObject: {fileID: 1357970550}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 1.62, y: 1.58, z: 1.8900003}
-  m_LocalScale: {x: -46.63, y: -22.87, z: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 1.8900003}
+  m_LocalScale: {x: 82.7, y: -63.2, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
@@ -1984,7 +1984,7 @@ Transform:
   m_GameObject: {fileID: 1926354358}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0.63, y: 2.84, z: -54.59}
+  m_LocalPosition: {x: 1.6, y: 2.84, z: -30.5}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -2047,10 +2047,10 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   m_LightProjectile: {fileID: 6483078321381274704, guid: 05650b6fc4c00df49a2c6f8784e58bb9, type: 3}
-  m_range: 50
-  m_radius: 0.5
-  m_velocity: 0.5
-  m_ProjectileFov: 50
+  m_range: 30
+  m_radius: 1
+  m_velocity: 2
+  m_ProjectileFov: 30
 --- !u!1001 &2040415378
 PrefabInstance:
   m_ObjectHideFlags: 0
